Renamed folder, moved variable
