+++ lib/translationModule/TranslationModule.js
@@ -2,9 +2,9 @@ import languages from './languages/collector';
 
 export default function translationModule() {}
 
-var language;
-
 translationModule.prototype.applyTranslation = function () {
+  var language;
+
   if (typeof window.f4a !== 'undefined') {
     language = languages[window.f4a.language];
   }
