+++ lib/translationModule/languages/collector.js
